requirements.txt
Django
numpy
torch
pandas
Pillow
fastcore
torchvision
torchmetrics
pytorch_lightning